Web Client (Mono and Flux)

+++ Spring-11-Rest-Jeckson/src/main/java/com/jackson/controller/WebFluxController.java
@@ -99,8 +99,21 @@ public class WebFluxController {
                 .bodyToMono(Genre.class);
     }
 
+//    @DeleteMapping("/delete/{id}")
+//    public Mono<Void> deleteWebClient(@PathVariable("id") Long id) {
+//        return webClient.delete()
+//                .uri("/delete-genre/{id}", id)
+//                .retrieve()
+//                .bodyToMono(Void.class);
+//    }
+
     @DeleteMapping("/delete/{id}")
-    public Mono<Void> deleteWebClient(@PathVariable("id") Long id) {
+    public Mono<Void> deleteWebClient(@PathVariable("id") Long id) throws Exception {
+        Integer countGenres = genreRepository.countGenresNativeQuery(id);
+        if (countGenres > 0) {
+            throw new Exception("Genre can't be deleted, is linked by a movie");
+        }
+
         return webClient.delete()
                 .uri("/delete-genre/{id}", id)
                 .retrieve()
